Revert "Move chat SSE to a dedicated Cloud Function (no Fastify in the response path)"
This reverts commit 7619099bdc7b9ae5a734280a2e680073d5ca6b5f.
