+++ functions/src/index.ts
@@ -1,5 +1,4 @@
 import { fastifyFunction } from './api'
-import { chatStream } from './api/chatStreamFunction'
 import { serviceApi } from './serviceApi/serviceApi'
 
-export { fastifyFunction as api, chatStream, serviceApi }
+export { fastifyFunction as api, serviceApi }
